feat: use a single error types file
