+++ src/errors/types.ts
@@ -11,3 +11,17 @@ export class RelayError extends Error {
     this.servicer_node = servicer_node
   }
 }
+
+/**
+ * @class LimitError
+ */
+export class LimitError extends Error {
+  method: string
+  message: string
+  constructor(message: string, method: string) {
+    super()
+    this.name = 'LimitError'
+    this.message = message
+    this.method = method
+  }
+}
