+++ src/services/pocket-relayer.ts
@@ -8,8 +8,7 @@ import { PocketAAT, Session, RelayResponse, Pocket, Configuration, HTTPMethod, N
 import { Redis } from 'ioredis'
 import { BlockchainsRepository } from '../repositories'
 import { Applications } from '../models'
-import { RelayError } from '../errors/relay-error'
-import { LimitError } from '../errors/limit-error'
+import { RelayError, LimitError } from '../errors/types'
 import AatPlans from '../config/aat-plans.json'
 import { blockHexToDecimal, checkEnforcementJSON } from '../utils'
